+++ admin/users.php
@@ -1,8 +1,8 @@
 <?php
 
 /*
-Version:     6.29
-Date:        12/01/26
+Version:     6.34
+Date:        02/02/26
 Name:        users.php
 Purpose:     User administrative tasks
 Notes:       {none}
@@ -93,7 +93,11 @@ $siteTitleEsc = htmlspecialchars($siteTitle, ENT_QUOTES, 'UTF-8');
     <link rel="manifest" href="/manifest.json" />
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link rel="stylesheet" type="text/css" href="/css/style<?php echo $cssver?>.css">
+    <link
+        rel="stylesheet"
+        type="text/css"
+        href="/css/style<?php echo $cssver?>.css?v=<?php echo $serviceWorkerVersion; ?>"
+    >
     <?php include APP_ROOT . '/includes/googlefonts.php';?>
     <script src="../js/jquery.js"></script>
     <script type="text/javascript">