+++ help.php
@@ -1,8 +1,8 @@
 <?php
 
 /*
-Version:     2.29
-Date:        12/01/26
+Version:     2.34
+Date:        02/02/26
 Name:        help.php
 Purpose:     Provides a help submission form and place for help notes.
 Notes:       {none}
@@ -42,7 +42,11 @@ $siteTitleEsc = htmlspecialchars($siteTitle, ENT_QUOTES, 'UTF-8');
     <meta name="viewport" content="initial-scale=1">
     <title><?php echo $siteTitleEsc;?> - help</title>
     <link rel="manifest" href="/manifest.json" />
-    <link rel="stylesheet" type="text/css" href="css/style<?php echo $cssver?>.css">
+    <link
+        rel="stylesheet"
+        type="text/css"
+        href="css/style<?php echo $cssver?>.css?v=<?php echo $serviceWorkerVersion; ?>"
+    >
     <?php include APP_ROOT . '/includes/googlefonts.php';?>
     <script src="/js/jquery.js?v=<?php echo $serviceWorkerVersion; ?>"></script>
 </head>