+++ issues.php
@@ -1,8 +1,8 @@
 <?php
 
 /*
-Version:     1.23
-Date:        13/01/26
+Version:     1.28
+Date:        02/02/26
 Name:        issues.php
 Purpose:     Issues page.
 Notes:       No db functions.
@@ -37,7 +37,11 @@ $siteTitleEsc = htmlspecialchars($siteTitle, ENT_QUOTES, 'UTF-8');
     <meta name="viewport" content="initial-scale=1">
     <title><?php echo $siteTitleEsc;?> - issues</title>
     <link rel="manifest" href="/manifest.json" />
-    <link rel="stylesheet" type="text/css" href="css/style<?php echo $cssver?>.css">
+    <link
+        rel="stylesheet"
+        type="text/css"
+        href="css/style<?php echo $cssver?>.css?v=<?php echo $serviceWorkerVersion; ?>"
+    >
     <?php include APP_ROOT . '/includes/googlefonts.php';?>
     <script src="/js/jquery.js?v=<?php echo $serviceWorkerVersion; ?>"></script>
 </head>