+++ reset.php
@@ -1,8 +1,8 @@
 <?php
 
 /*
-Version:     3.37
-Date:        13/01/26
+Version:     3.42
+Date:        02/02/26
 Name:        reset.php
 Purpose:     Password reset page, called from login.php.
 Notes:       Not a secure page!
@@ -233,7 +233,8 @@ $siteTitleEsc = htmlspecialchars($siteTitle, ENT_QUOTES, 'UTF-8');
     <link
         rel="stylesheet"
         type="text/css"
-        href="css/style<?php echo htmlspecialchars($cssver, ENT_QUOTES, 'UTF-8');?>.css"
+        href="css/style<?php echo htmlspecialchars($cssver, ENT_QUOTES, 'UTF-8');?>.css?v=<?php
+        echo $serviceWorkerVersion; ?>"
     >
     <?php include APP_ROOT . '/includes/googlefonts.php';?>
 </head>